Fix DNS server alignment once and for all...*
Lmao, found this bug when I was adding NA and EU Images to represent the regions, for some reason the EU side wasn't aligning up properly. Spent 10 minutes only to find out I made a class that fixed another issue (I think) specifically for North American side. As before only had 1 DNS server. So it was obsolete code once EU was added. Never removed it.

This patch adds the class, which I have changed the name to alignfix. Which fixes this issue, and properly aligns both sides now.

Thanks older me. This is becoming worse and worse to work with lol.